tout marche
manque le top_k de maxime

+++ gidi.py
@@ -118,7 +118,6 @@ def index_local():
         cur1 = conn.cursor()
         cur2 = conn.cursor()
 
-        print fmaladie_name
         with open(fmaladie_name) as f:
             maladies = (f.readlines())
 
@@ -162,9 +161,15 @@ def index_global():
                 "DROP TABLE IF EXISTS index_global;"
                 "CREATE TABLE index_global(adress CHAR(100), proba NUMERIC(10,8));"
             )
+            tab_ip = []
             sick = request.form['maladie']
+            ip1 = request.form['ip1']
+            tab_ip.append(ip1)
+            ip2 = request.form['ip2']
+            tab_ip.append(ip2)
+            ip3 = request.form['ip3']
+            tab_ip.append(ip3)
             doOrNo = request.form['choice']
-            print sick
             cur2.execute(
                 #creation de table_path et la remplir
                 'DROP TABLE IF EXISTS table_path;'
@@ -175,30 +180,23 @@ def index_global():
                     CREATE TABLE table_path_ip(adress CHAR(100));"""
             )
 
-            print '++++++++++++++++++++++++++'
             cur2.execute(
                 """SELECT adress FROM table_path"""
             )
             listeFarms = cur2.fetchall()
             for farm in listeFarms:
-                print farm[0]
                 ipName = session['host'] + '-' + farm[0]
-                print ipName
                 cur2.execute(
                     """
                     INSERT INTO table_path_ip(adress) VALUES('""" + ipName + """')"""
                 )
-            print '++++++++++++++++++++++++++'
 
             cur3.execute(
                 #pour remplir la listeTablesSick en bas
                 'SELECT adress FROM table_path;'
             )
             listeTablesSick = cur3.fetchall()
-            print listeTablesSick
             for liste in listeTablesSick:
-                print '-----------------------'
-                print liste[0]
                 cur3.execute(
                     #remplir l'index global
                     """INSERT INTO index_global(adress, proba)
@@ -206,8 +204,6 @@ def index_global():
                     DELETE FROM table_path WHERE ctid IN (SELECT ctid FROM table_path LIMIT 1);
                     DELETE FROM table_path_ip WHERE ctid IN (SELECT ctid FROM table_path_ip LIMIT 1)"""
                 )
-            print 'en haut pas touche'
-            test = 'toto3'
             cur4.execute(
                 # mettre en ordre l'index global
                 """CREATE TABLE test (LIKE index_global);
@@ -222,25 +218,27 @@ def index_global():
                 DROP TABLE IF EXISTS table_path_ip;
                 """
             )
-            print doOrNo + '==========================================='
             if (doOrNo == '1'):
-                # dblinK
-                cur4.execute(
-                    """
-                        SELECT dblink_connect('toto6', 'hostaddr=192.168.43.143 port=5432 dbname=PPD user=postgres password=root');
-                        INSERT INTO index_global
-                        SELECT * FROM dblink('toto6','SELECT * FROM index_global') AS t(a text ,c numeric );
-                        CREATE TABLE test (LIKE index_global);
-                        INSERT INTO test
-                        SELECT * FROM index_global ORDER BY proba DESC;
-                        DROP TABLE index_global;
-                        ALTER TABLE test RENAME TO index_global;
-                        SELECT dblink_disconnect('toto6');
-                      """
-                )
-            else:
-                print 'bip bip chui pas rentré'
-            print 'okkkkkkkkkkkkkkk'
+                for ip in tab_ip:
+                    print ip
+                    if ip:
+                        # dblinK
+                        cur4.execute(
+                            """
+                              SELECT dblink_connect('conx', 'hostaddr="""+ip+""" port=5432 dbname=PPD user=postgres password=root');
+                              INSERT INTO index_global
+                              SELECT * FROM dblink('conx','SELECT * FROM index_global') AS t(a text ,c numeric );
+                              CREATE TABLE test (LIKE index_global);
+                              INSERT INTO test
+                              SELECT * FROM index_global ORDER BY proba DESC;
+                              DROP TABLE index_global;
+                              ALTER TABLE test RENAME TO index_global;
+                              SELECT dblink_disconnect('conx');
+
+                            """
+                        )
+                    else :
+                        print "pas de ip rentrer"
             conn.commit()
             return render_template('index_global_confirm.html', active="index_global")
         return render_template('index_global.html', active="index_global")
@@ -258,8 +256,8 @@ def threshold():
             conn = connect()
             cur1 = conn.cursor()
             cur2 = conn.cursor()
-            seuil = request.form['seuil']
-            print seuil + '==============================='
+            seuil_check = request.form['seuil']
+
             cur1.execute(
                 """
                 DROP TABLE IF EXISTS threshold;
@@ -269,7 +267,7 @@ def threshold():
                 DROP TABLE IF EXISTS index_global_bis;
                 CREATE TABLE index_global_bis (LIKE index_global);
                 INSERT INTO index_global_bis
-                SELECT * FROM  index_global WHERE proba>= """+seuil+""";
+                SELECT * FROM  index_global WHERE proba>= """+seuil_check+""";
                 """
             )
             cur1.execute(
@@ -284,15 +282,13 @@ def threshold():
                 tabAdressFull = adressFull.split("-")
                 ip = tabAdressFull[0]
                 farm = tabAdressFull[1]
-                print ip
-                print farm
                 if (ip == 'localhost'):
                     cur2.execute(
                         """
                         INSERT INTO help_proba
                         SELECT proba FROM """+farm+"""
                         WHERE proba>=
-                        """+seuil+"""
+                        """+seuil_check+"""
                         """
                     )
                     cur2.execute(
@@ -317,7 +313,7 @@ def threshold():
                     cur2.execute(
                         """
                             INSERT INTO help_proba
-                            SELECT * FROM dblink('conx','SELECT proba FROM """ + farm + """ WHERE proba>=""" + seuil + """ ') AS t(c numeric );
+                            SELECT * FROM dblink('conx','SELECT proba FROM """ + farm + """ WHERE proba>=""" + seuil_check + """ ') AS t(c numeric );
                         """
                     )
 
@@ -335,8 +331,15 @@ def threshold():
                         SELECT dblink_disconnect('conx')
                         """
                     )
-
                 tab.sort()
+            cur1.execute(
+                """
+                DROP TABLE IF EXISTS threshold;
+                DROP TABLE IF EXISTS help_proba;
+                DROP TABLE IF EXISTS index_global_bis;
+                """
+            )
+
             conn.commit()
             return render_template("threshold_result.html", active="threshold", res=tab)
         return render_template("threshold.html", active="threshold")
